==> Analysis_code/06_model_based.Rmd <==
---
title: "Hybrid model-based HRS (integrating patient covariates)"
subtitle: "Mboost: model-based boosting" 
author: "Hannah Marchi"
date: ' `r format(Sys.Date(), "%d\\.%m\\.%Y")`'
output:
  html_document:
    toc: true
    toc_depth: 4
    toc_float: true
    number_sections: true
    df_print: paged
    code_folding: hide
  pdf_document:
    toc: true
    toc_depth: '4'
editor_options:
  chunk_output_type: console
always_allow_html: true
params: 
  initialValues: "memory1" # memory1, memory2
  matrixUpdate: "iterative" # direct, iterative
  iterationStop: 10
  patientCovars: TRUE
  testStrategy: "RS" # classic (row-wise), RS (cell-wise)
  therapyChoice: "single" # single, combi
  fold: 10
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE)
```

```{r}
#load packages
library(dplyr)
library(rlist)
library(caret)
library(pROC)
library(stringr)
```

```{r}
source("03_HRS_functions.R")

# single/combi therapies
if(params$therapyChoice == "single"){
  load("HRS_data_single.Rdata")
} else if(params$therapyChoice == "combi"){
  load("HRS_data_combi.Rdata")
}
#-----
# include/exclude patient covariates
if(params$patientCovars == TRUE){
  myData_model <- myData_scale
} else {
  myData_model <- myData_onlyTherapies
}
```

# Setup

The following parameters are used for analysis:  

* initialValues:  `r params$initialValues`    
* matrixUpdate: `r params$matrixUpdate`    
* iterationStop: `r params$iterationStop`      
* patientCovars: `r params$patientCovars`   
* trainingTestStructure: `r params$testStrategy`  
* therapyChoice: `r params$therapyChoice`   
* CV fold: `r params$fold`  


## Application

### similarity threshold for patients

```{r}
cutoff_patient <- 0.8
cutoff_therapy <-  0.90
nfold <- params$fold

if(params$testStrategy == "RS"){
  RS_modelBased_th <- validate_RS(HRS = "model",
                                  input_data = myData_model,
                                  therapies = therapies,
                                  seed = 202411,
                                  patientSim = simP,
                                  therapySim = simT,
                                  k = NULL,
                                  thresholdP = cutoff_patient,
                                  simChoice = "threshold",
                                  thresholdT = cutoff_therapy,
                                  stop_iter = params$iterationStop,
                                  initialValues = params$initialValues,
                                  matrixUpdate = params$matrixUpdate, 
                                  fold = nfold) 
} else if(params$testStrategy == "classic"){
  RS_modelBased_th <- validate_classic(HRS = "model",
                                       input_data = myData_model,
                                       therapies = therapies,
                                       seed = 202411,
                                       patientSim = simP,
                                       therapySim = simT,
                                       k = NULL,
                                       thresholdP = cutoff_patient,
                                       simChoice = "threshold",
                                       thresholdT = cutoff_therapy,
                                       stop_iter = params$iterationStop,
                                       initialValues = params$initialValues,
                                       matrixUpdate = params$matrixUpdate, 
                                       fold = nfold) 
}
```

```{r}
bin_cutoff <- 0.5

# overall validation measure over all folds
RS_modelBased_th$df_results$compare$y_true <- as.factor(RS_modelBased_th$df_results$compare$y_true)
myROC <- roc(response = RS_modelBased_th$df_results$compare$y_true, predictor = RS_modelBased_th$df_results$compare$y_pred)
RS_modelBased_th$df_results$compare$pred_bin <- ifelse(RS_modelBased_th$df_results$compare$y_pred >= bin_cutoff, 1, 0) 
conf <- confusionMatrix(factor(RS_modelBased_th$df_results$compare$pred_bin), factor(RS_modelBased_th$df_results$compare$y_true), positive = "1")
results_RS_modelBased_th <- c("AUC" = round(as.numeric(myROC$auc), 3), #auc
                     conf$overall[1], # accuracy
                     conf$byClass[c(1,2,7,5)]) # sens, spec, F1, Prec

# validation measures for single folds
perfFolds <- matrix(nrow = nfold, ncol = 6)
colnames(perfFolds) <- c("AUC", "Accuracy", "Sensitivity", "Specificity", "F1", "Precision")  

for(j in 1:nfold){
  data_compare <- RS_modelBased_th[["results"]][[j]][["compare"]]
  
  data_compare$y_true <- as.factor(data_compare$y_true)
  myROC <- roc(response = data_compare$y_true, predictor = data_compare$y_pred, quiet = TRUE)
  data_compare$pred_bin <- ifelse(data_compare$y_pred >= bin_cutoff, 1, 0) 
  conf <- confusionMatrix(factor(data_compare$pred_bin), factor(data_compare$y_true), positive = "1")
  perfFolds[j,] <- c(round(as.numeric(myROC$auc), 3), #auc
                     conf$overall[1], # accuracy
                     conf$byClass[c(1,2,7,5)]) # sens, spec, F1, prec
}
perfFolds <- perfFolds[,c(1,3,4,6,5)]
```

```{r}
plot(myROC, main = paste0("ROC: Model-based, patientThr_", cutoff_patient, "_therapyThr_", cutoff_therapy))
print(paste0("Overall performance over all folds for P_ ", cutoff_patient, "_T_", cutoff_therapy, " with binary cutoff ", bin_cutoff, ":"))
print(results_RS_modelBased_th)

print(paste0("Summary of performance for all folds for P_ ", cutoff_patient, "_T_", cutoff_therapy, " with binary cutoff ", bin_cutoff, ":"))
print(apply(perfFolds, 2, summary))
```


```{r}
sessionInfo()
save.image(paste0("workspace_II_", paste0(unlist(params), collapse = "_"), ".Rdata"))
```
